requirements-dev.txt
-r requirements.txt

attrs==22.2.0
cfgv==3.3.1
distlib==0.3.6
django-debug-toolbar==4.0.0
filelock==3.10.7
exceptiongroup==1.1.1
flake8==6.0.0
identify==2.5.22
iniconfig==2.0.0
mccabe==0.7.0
mypy==1.1.1
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.0
platformdirs==3.2.0 
pluggy==1.0.0
pre-commit==3.2.1
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.2.2
pyyaml==6.0
tomli==2.0.1
typing_extensions==4.5.0
virtualenv==20.21.0
